fix: gate tunnel writes on a serving stream, let a reclaimed instance re-enrol
Gateway: Connect registers a stream before serve starts its pumps, so a request
could be queued in that window; a caller that gave up there was told "cancelled
before dispatch, retryable" while the head sat in the outbound queue and shipped
the moment the pump started — a retry repeated the side effect and the late
response for an unknown request ID tore the stream down. enqueue now waits on a
serving gate that serve closes, so nothing enters the queue before the pump
exists and dispatched alone decides before/after.

Enrollment keys: an instance record is reclaimed 30 days after its last contact,
so a host that came back later presented a key for a record that no longer
existed, got 401 invalid_key (terminal, no bootstrap fallback by design) and
restart-looped. The API now answers a well-formed key for a missing record with
401 instance_unknown — never a revocation, since tombstones are excluded from
the TTL — and the connector re-enrols with its bootstrap token on exactly that
code, or exits 2 naming the fix when no token is configured.

Lint: CI pins golangci-lint v2.9.0, whose nolintlint rejects a //nolint:gosec on
a line its older gosec does not flag (G704 does not exist there) and whose gosec
flags a uint(attempt) shift the newer one accepts. Converted the four
version-sensitive directives to #nosec and dropped the conversion; verified with
the pinned version.

+++ internal/connect/manager_test.go
@@ -235,7 +235,7 @@ func run(t *testing.T, m *Manager) (stop context.CancelFunc, result chan error) 
 func TestBackoffIsCappedAndJittered(t *testing.T) {
 	const base, capped = time.Second, 30 * time.Second
 	for attempt := range 10 {
-		want := min(base<<uint(attempt), capped)
+		want := min(base<<attempt, capped)
 		if attempt > 30 { // guard against shift overflow in the test itself
 			want = capped
 		}
